+++ modules/fortigate/fgt_asg_with_function/variables.tf
@@ -201,8 +201,7 @@ variable "cloud_function" {
     license_source      = optional(string, "none")
     license_file_folder = optional(string, "./licenses")
     autoscale_psksecret = optional(string, "psksecret")
-    print_debug_msg     = optional(bool, false) # Deprecated, use logging_level instead
-    logging_level       = optional(string, "NOT_SPECIFIED")
+    logging_level       = optional(string, "NONE")
     fortiflex = optional(object({
       retrieve_mode = optional(string, "use_stopped")
       username      = optional(string, "")
@@ -235,9 +234,7 @@ variable "cloud_function" {
             "file_fortiflex" : Injecting licenses based on license files first. If all license files are in use, try FortiFlex next.
         - license_file_folder : (Optional | string | default:"./licenses") The folder where all ".lic" license files are located. Default is "./licenses" folder.
         - autoscale_psksecret : (Optional | string | default:"") The secret key used to synchronize information between FortiGates. If not set, the module will randomly generate a 16-character secret key.
-        - print_debug_msg : (Optional | bool | default:false) Deprecated, use logging_level instead. If set to true, the cloud function will print debug messages. You can find these messages in Google Cloud Logs Explorer.
-        - logging_level : (Optional | string | default:"NOT_SPECIFIED") Verbosity of logs. Possible values include "NONE", "ERROR", "WARN", "INFO", "DEBUG", and "TRACE". You can find logs in Google Cloud Logs Explorer.
-        For backward compatibility reasons, the default value "NOT_SPECIFIED" functions the same as "NONE" and logs nothing unless you set the deprecated "print_debug_msg" to true, in which case it acts like "INFO".
+        - logging_level : (Optional | string | default:"NONE") Verbosity of logs. Possible values include "NONE", "ERROR", "WARN", "INFO", "DEBUG", and "TRACE". You can find logs in Google Cloud Logs Explorer.
         - fortiflex : (Optional | object) You need to specify this parameter if your license_source is "fortiflex" or "file_fortiflex".
             - retrieve_mode : (Optional | string | default:"use_stopped") How to retrieve an existing fortiflex license (entitlement):
                 "use_stopped" selects and reactivates a stopped entitlement where the description field is empty;
@@ -281,10 +278,10 @@ variable "cloud_function" {
 # AutoScaler
 variable "autoscaler" {
   type = object({
-    max_instances   = number
-    min_instances   = optional(number, 2)
-    cooldown_period = optional(number, 300)
-    cpu_utilization = optional(number, 0.9)
+    max_instances        = number
+    min_instances        = optional(number, 2)
+    cooldown_period      = optional(number, 300)
+    cpu_utilization      = optional(number, 0.9)
     autohealing = optional(object({
       health_check_port   = optional(number, 8008)
       timeout_sec         = optional(number, 5)
@@ -292,6 +289,7 @@ variable "autoscaler" {
       unhealthy_threshold = optional(number, 10)
       }), {}
     )
+    scale_in_control_sec = optional(number, 300)
   })
   description = <<-EOF
     Auto Scaler parameters. This variable controls when to autoscale and the maximum number of instances.
@@ -306,6 +304,7 @@ variable "autoscaler" {
             - timeout_sec         : (Optional | number | default:5) How long (in seconds) to wait before claiming a health check failure.
             - check_interval_sec  : (Optional | number | default:30) How often (in seconds) to send a health check.
             - unhealthy_threshold : (Optional | number | default:10) A so-far healthy instance will be marked unhealthy after this many consecutive failures.
+        - scale_in_control_sec : (Optional | number | default:300)  When the group scales down, Google Cloud will delete at most one FGT every 'scale_in_control_sec' seconds.
 
     Example:
     ```
@@ -314,6 +313,7 @@ variable "autoscaler" {
         min_instances = 2
         cooldown_period = 300
         cpu_utilization = 0.9
+        scale_in_control_sec = 300
     }
     ```
     EOF
@@ -323,4 +323,15 @@ variable "config_script" {
   type        = string
   default     = ""
   description = "Extra config data"
+}
+
+variable "special_behavior" {
+  type = object({
+    disable_secret_manager  = optional(bool, false)
+  })
+  default = {}
+  description = <<-EOF
+    This variable can specify special behavior to suit various needs.
+    Do not use this variable unless instructed by the author.
+  EOF
 }